requirements.txt
crewai==0.28.8
streamlit==1.29.0
python-dotenv==1.0.0
openai>=1.13.3
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-anthropic>=0.1.0
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.0
supabase>=2.0.0
psycopg2-binary>=2.9.0
langchain>=0.1.0
langchain-text-splitters>=1.0.0
numpy>=1.24.0
pypdf>=3.17.0
python-docx>=1.1.0
anthropic>=0.25.0
google-generativeai>=0.3.0
langchain-google-genai>=1.0.0

